[FIX] account: popover payment information unclickable
In the account invoice the payments are shown and a popover with more
information is displayed when clicking on an (i) icon.

In this popover two buttons are available, but the used type of popover
disappear immediately when clicking on them, thus the buttons were
directly removed and would not work.

This commit add a delay when hiding the popover so the buttons
functionnality is not lost.

opw-680095