[FIX] point_of_sale: display the correct UOM in the ScaleScreenWidget
The scale screen widget was hardcoded to show kilograms as the
UOM. This makes it so that the UOM set on the product being weighed is
used instead.

Note that this bug was purely visual. It did not convert the weight to
the wrong UOM or anything like that.

See also 0541f49542481043b1fb49f7842fcc2249f40e2b.

opw-684374