[FIX] calendar: use correct bootstrap for invitations
Commit b3fdee844510f636923f0e8d14b78e444c81bbbe introduced a new template for meeting invitation page.
This template used bootstrap row's without any container which added
an ugly useless scrollbar at the bottom.

This commit uses valid bootstrap by simply removing the bootstrap
column layout (as there is no column but only centered content) and
using a simple container. Also, wraps the table in a .table-responsive
div so that the template is mobile-friendly.
Also fixes a typo in the template name.