[FIX] base: set current company as default for currency rates
When manually creating/importing a currency rate, there is no default
on the company_id field.

This is quite unfortunate, as rates are filtered by company then by
date, which means that a rate without company will never be taken
into account despite being newer. This filter makes sense for multi
company setups, but for mono-company setups this is quite annoying.

This commit adds a default on the company_id and adds a view for
the base_currency_rate model to only display the company field in
multi company settings.