[FIX] report: Pass correctly the docids to render_html method
In the 'get_html' method, the report is using a custom model to render its html,
we should use its own 'render_html' method to generate the report. During the
module migration, we replaced
'report_model.render_html(cr, uid, ids, data=data, context=context)'
by
'report_model.render_html(data=data)'

The new API signature of 'get_html' is

    @api.model
    def get_html(self, docids, report_name, data=None):

That way, 'render_html' doesn't receive docids anymore when called. We should do instead

    if report_model is not None:
        return report_model.browse(docids).render_html(data=data)

As render_html is an api.multi method, this is all wrong too, because we'll send a
browse record with the custom report model to the render_html method, to use the
ids with 'self.ids' that are absolutely not corresponding to the report model but to
to the record model we want to print (an invoice for example).

In old API style we were sending a list of ids, but it's not correct to keep the same
behavior by sending a browse record with only the ids relevant. We have also to
change render_html method decorator to @api.model with a new signature

    @api.model
    def render_html(self, docids, data=None):

And replace the call to render_html by

    if report_model is not None:
        return report_model.render_html(docids, data=data)

In the case the report is printed from a wizard, this is sufficient because the ids
are taken from the active_ids in context, or in data.

In the case the report is printed from a simple action, we should replace the occurence
of 'self.ids' (which doesn't mean anything) to 'docids'

This fix doesn't solve the printing issues for the custom accounting reports, because
the method get_pdf has been reimplemented in the module account_reports.