[FIX] web_editor: paragraph dropdown style
Before commit 853bc426e4d2f03bdeeeac8cfe42a81ee985770a, the paragraph alignment dropdown menu was
too large for its content, so it was ugly. After that commit, it became
too small for its content, which is even more ugly.

This fix sets the width of this specific dropdown so that its content
fills exactly in (as with original summernote).