Revert "[IMP] project: tour mini improvement"
This reverts commit 8dc2ad1adebaf1ba03a2c56cf3920ddabdcf7089.

The commit failed the tour test since there is invisible
".o_kanban_ghost" cards after the last. Since the created new
project is already sorted in the right location, we don't know
if it will be the first or the last.

But if it is the first, we know we will see the help tip.