[FIX] web_tour: properly filter on visible elements
When checking the DOM for tour step trigger, the ":visible" jQuery
filter was used to get the visible elements. The problem is that this
filter does not care about the visibility css property. So, for jQuery,
a display:block and visibility:hidden element is visible.

This was an issue as kanban ghosts or settings panels in odoo are
hidden thanks to the visibility property instead of display.