[FIX] purchase: fiscal position in multi-company environments
Properties depends on with which companies the records
are browsed.

When retrieving the fiscal position as sudo (from a cron), we should force
the company from the context.

After https://github.com/odoo/odoo/commit/f1c2845 as the context is force in get_fiscal_position,
we should call this method with a company (and not pass the company in the context).
This has been forgot in https://github.com/odoo/odoo/commit/fb8fa21

opw-684802