[FIX] website: avoid redirect to website after login
When logged in a user was redirected to the website instead of the backend.
This was due to a wrong group verification, at that point:
- request.uid = ID of the user authenticating (e.g. admin)
- request.env.uid = ID of the user that rendered the page (public)

As public does not belong to `base.group_user` group, the user was redirected to
the website.

Introduced at 283a5c8d during migration.

Fixes #13203
Closes #13204