[FIX] base: improve last chance for detection of mimetype
Since we return nosniff header, we need to avoid the magic mimetype.

In case of picture in compute field (eg. image_small on product),there are
no mimetype saved in column, and impossible to detect it from name/extention
because url are like: /web/image/product.product/47/image_small.

So before to arbitrary return default_mimetype, we try to detect it from the
content itself.

This commit continue to fix odoo/odoo@13193