[IMP] website_blog: use valid website snippet as default content for blog post
Adapt the default content of blog post to be a valid website snippet.
Previous content was not entirely possible to remove (there still was
a remaining .container) and was not easily customizable (user had to
remove it if he wants something great).

Note: maybe we should adapt the snippets so that they are callable with
custom content or do something like what has been done for default
content for job positions (see commit 28bfaee2e3fb77d4b1190694da317b4d103be5cc).